Add better logging and use the command structure
This change adds more logging to the server and client applications
when debug is enabled. This also allows the client to follow the
job command model that we now have due to the message specification
being unified.

The command structure made it possible to improve command routing
and more intellegent cache hit and miss functionality.

More options have been documented in the README.

Signed-off-by: Kevin Carter <kecarter@redhat.com>