Add an option UI component that can start with the server
This change implements a simple read-only UI that can query active in memory
jobs and display details about their execution. The UI is implemented in flask
and while flask is installed with the package, the execution of the flask
components is totally optional. New command line server arguments have been
added to support this usecase.

Signed-off-by: Kevin Carter <kecarter@redhat.com>