Implements new orchestrations improves speed
This change implements a new set of orchestration examples which
highlight how complex operations can be accomplished within a
cluster. The new orchestartion is for RabbitMQ and implements a
fully SSL and clustered rabbitmq environment.

Changes have been made to the server and client code to allow this
new orchestration to run efficiently and provide robust information
on the execution when required.

Another fix that went in was really a simplification which changes
the orchestration code to load all members of the targets at once,
thereby giving the job just one ID and an in-sync execution. This
cuts down on cycles especially when dealing with large clusters
and complex orchestrations.

Signed-off-by: Kevin Carter <kecarter@redhat.com>