Begin working on async operations (#119)
* [WIP] Add async job processing

This change adds an optional argument to the orchestration function
which will allow a task to run in an async setup. When enabled, parent
jobs are grouped into queues which are then processed at a maximum of 5
parents jobs at time. This gives operators the ability to further
improve their delivery times by taking control of the execution scheme.

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* Add management function to dump the local disk cache

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* Update complex arg support

This change allows the server to also inspect and adjust args when
setting them in cache.

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* update unique parents

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* fix rotating cache eviction issues

Signed-off-by: Kevin Carter <kecarter@redhat.com>