Re-work query and implement server callbacks
This change re-works the query function to better support the new async
capabilities within Directord. To support this option components have
the ability to generate follow-on tasks which cause an orchestration to
block until the follow-on task has completed.

This change cleans-up the query hacks we had in server and client code
by using new job generation instead of attempting to run through a
series of string matching tomfoolery.

Signed-off-by: Kevin Carter <kecarter@redhat.com>