remove the use of libssh2 (#138)
* Remove the use of libssh2

Upon review the use of libssh2 has been found to be not FIPS compatible.
This change moves directord to libssh which is FIPS compatible. At
present, the use of libssh2 makes it impossible to connect to a FIPS
enabled system, which is effectively useless for our purposes. This
change moves the lib accordingly and ensures we're building the most
FIPS compatible solution from day 0.

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* move bootstrap to a new module

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* use host specific libssh session

Signed-off-by: Kevin Carter <kecarter@redhat.com>