Fix set_cache to use get and not pop
There can be a race condition when updating a value because pop would
remove it from the cache for a period of time. Switching to get will
prevent the value from completely disappearing if others attempt to
fetch it while it's being updated. There is still a race if two tasks
are running in parallel that expect a value to be provided by one of the
tasks for the other.