Improve async task tracing
Async tasks will now use the parent ID within the queue name. This is
done so that operators can now see when a queue is executing and when it
is pruned.

Allow query to run without waiting for the cache to be populated with
the queried value.

All exec tasks can now be forced into an async mode.

Signed-off-by: Kevin Carter <kecarter@redhat.com>