Rework the transfer protocol
The transfer protocol will now use an explicit flow control ensuring
messages never overwhelm the system, are well regluated on both the
server and client, and more intellegent.

The transfer component will now make requests of the server for specific
chunks of data instead of allowing the server to send an unlimited set
of chunks. While this may have a performance impact when compared to a
serial processing of large files, it will ensure consistency throughtout
the process.

Signed-off-by: Kevin Carter <kecarter@redhat.com>