Add support for oslo-messaging as a driver (#213)
* Add interface to driver class

The driver needs a reference to the interface class (Client/Server), in
order to implement interface specific callbacks (handle_heartbeat,
handle_job, etc), and have access to data associated with the interface
classes (bind_address, workers, etc).

Signed-off-by: James Slagle <jslagle@redhat.com>

* Add oslo-messaging driver stub

Adds an initial driver stub to use with oslo-messaging

Signed-off-by: James Slagle <jslagle@redhat.com>

* Add run method to driver and implement for oslo-messaging

Adds a run method which is called in its own thread in each interface
class. Not all drivers will need to implement run, but in the case of
oslo-messaging, we need to start qdrouterd.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Abstract heartbeat and add support to oslo-messaging

Abstract the heartbeat handling in order to add support for
oslo-messaging. The heartbeat handling is no longer ZeroMQ specific. The
following changes were done to make the heartbeat handling more abstract
for other drivers:

* The "bind_job" socket is now stored on the driver object (zmq) instead
  of the interface object (client/server).
* The driver now takes care of encoding/decoding RPC arguments (data
  dictionary / identity) for the heartbeat, as not all drivers require
  encoding.
* A uuid attribute is added to the interface, which is needed for the
  oslo-messaging driver.
* The heartbeat logic is now refactored into a "handle_heartbeat" method
  on the Server class, to make it callable by the oslo-messaging driver.

Signed-off-by: James Slagle <jslagle@redhat.com>