add hostname fencing (#228)
While it all hosts should have unique names, it is possible for two
machines to be misconfigured, resulting in multiple machines with the
same host name. To ensure we're able to address machine correctly and
that the server is maintaining a correct inventory, the heartbeat
process will now collect the machine id and use that unique id for
fencing. Should multiple machines with the same hostname but different
machine ids try to join the cluster an error will be raised and the new
node will not be permitted to join. A log entry will be generated on
how to fix the issue.

Signed-off-by: Kevin Carter <kecarter@redhat.com>