fix bug 235
The messaging driver was bypassing the server and client process thread.
This change updates the messaging driver to ensure that the server and
client processing work within the driver abstract, which will ensure
that the rest of the application functions as expected.

The backend processes have been stubbed out but are not actually needed
given the messaging driver runs a persistent thread.

Fixes Issue #235

Signed-off-by: Kevin Carter <kecarter@redhat.com>