Generate certs for all clients centrally
Refactors the SSL setup to generate all the client certs on the server
first, then copy them out to each client individually. This results in
all certs being signed by the CA.

Also simplifies the certificate options in the messaging driver, as
separate certs for the server and client are not needed. In cases where
the server and client both run locally, they can use the same cert
without issue.

Also generates a unique certificate and key just for qdrouterd, also
signed by the same CA.

Signed-off-by: James Slagle <jslagle@redhat.com>