Add SSL support for messaging driver (#247)
* Add SSL support for messaging driver

Adds SSL support to the messaging driver and the bootstrap catalog for
the messaging driver. Also updates dev-setup.sh to only necessary
packages depending on which driver is chosen.

This also updates the Github workflows for the functional_messaging
tests to setup the needed SSL configuration.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Add arguments to messaging driver

Adds custom CLI arguments to the messaging driver and adds an
interface from main.py to load arguments from any drivers that expose
them.

A followup patch will refactor the zmq driver to use this same pattern.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Generate certs for all clients centrally

Refactors the SSL setup to generate all the client certs on the server
first, then copy them out to each client individually. This results in
all certs being signed by the CA.

Also simplifies the certificate options in the messaging driver, as
separate certs for the server and client are not needed. In cases where
the server and client both run locally, they can use the same cert
without issue.

Also generates a unique certificate and key just for qdrouterd, also
signed by the same CA.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Document SSL encryption for messaging driver

Signed-off-by: James Slagle <jslagle@redhat.com>