Add option to allow orchestrations to override targets
This change will allow operators to override the targets for a given set
of tasks. The new `assign` key provides the ability to define a list of
targets which may be outside the scope of the current orchestration.

The `assign` option is a list of all intended targets and is validate
upon orchestration ingest.

Closes #269

Signed-off-by: Kevin Carter <kecarter@redhat.com>