Fix messaging bootstrap for multiple nodes (#259)
* Honor env vars in dev-setup.sh

Also honor env vars set externally from the script instead of just
relying on positional arguments.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Add directord_bootstrap magic var for server bootstrap

This magic var needs to be available when bootstrap_run is used with the
server, otherwise blueprinting in the server tasks will not work.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Fix bind address in messaging driver

The server can use the local hostname (self.identity) as the bind
address for where to find qdrouterd, as for now we assume that qdrouterd
runs on the server node.

The client needs to honor the value set in the arg for --server-address.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Fix messaging bootstrap for multiple nodes

Previous, the messaging bootstrap and SSL setup was only working when
the server and client were the same node. These changes make it work for
when the server and client(s) are separated across multiple nodes.

This patch also changes the SSL setup to use openssl instead of
certmonger.

Signed-off-by: James Slagle <jslagle@redhat.com>

* Rename ZMQ server-address and bind-address options

server-address and bind-address are specific to the ZMQ driver, and do
not make sense for the messaging driver. This change renames the options
and associated configurations and environment variables to start with
zmq to match the other zmq specific options.

Adds a messaging specific option, --messaging-address for specifying the
address of the router/broker server when using the messaging driver.

Signed-off-by: James Slagle <jslagle@redhat.com>