Add option to allow operators to set the machine id
This change allows operators to define the machine id. While the machine
id should be auto generated, when running in containerized infra, the ID
is passed through to the container, so to be able to run multiple
directord containers on a single host, we need the ability to define a
unique id.

Signed-off-by: Kevin Carter <kecarter@redhat.com>