Job interaction improvements
Server side job handling was all done within a consistent loop, this
change adds a quick handling check when sending so that the server is
always processing data, and never in an idle state when work is needed
to be done.

The credit process is no longer needed and removing it will improve
performance by about 5%. This change removes the credit functions
from the core client and server and improves some logging.

Meta rev to 0.11.2

JOB ACK was removed. The job ack control no longer has any purpose given
the use of the processing control. The job ack was an old interaction
which was kept for legacy purposes. Removing this interaction improves
server throughput by removing processing overhead.

Signed-off-by: Kevin Carter <kecarter@redhat.com>