update file store (#300)
The filestore, depending on the file system, may return a differnet
create time than the actual time a file is created. This change
implements a builtin method into the Cache class which will return the
birthtime based on object attributes and fall back to stat when
required. This change will allow the Cache method to be more accurate,
when used as a datastore, and ensure efficiency.

Signed-off-by: Kevin Carter <kecarter@redhat.com>