use threading instead of multiprocessing
While the multiprocessing API should work for our purposes, the use of
multiprocessing can cause issues with different drivers. This change
modifies directord to make use of the threading library and native queue
functions instead of relying on multiprocessing, which should ensure it
is more compatible with different backends.

The driver code will now inform the server and client to ensure we're
running with the proper method (multiprocessing|threading).

Signed-off-by: Kevin Carter <kecarter@redhat.com>