ensure that drivers are used process based locks
Locks can not be mixed and matched, this change ensures that the driver
code, which is responcible for the process execution is selecting the
correct lock types.

To ensure everything is correct, we've also name spaced the queues
created. This will ensure that multiprocessing queues are only used when
required.

Signed-off-by: Kevin Carter <kecarter@redhat.com>