Name package directord
directord isn't a python library, so drop the python/python3 prefix.
Packaging requirements[1] also require that the package name match the spec
file name. As a result, the redundant summary and description sections
can be removed.

Also bumps the default release version to match that from
directord/meta.py and fixes some typos in the description.

[1] https://docs.fedoraproject.org/en-US/packaging-guidelines/#_spec_file_naming

Signed-off-by: James Slagle <jslagle@redhat.com>