use pre-fork signals to allow exit (#351)
This change will allow directord to attempt a graceful shutdown of the
application. Because signals can not be used in multi-threading, all
of the signal processing is handled pre-fork|thread. The server|client
classes then use the driver event handler to begin a system shutdown
when a signal is encountered.

Signed-off-by: Kevin Carter <kevin@cloudnull.com>