move key generation to the zmq driver
The key generation bits in directord were intended to be an abstraction,
however, they're only supported by the one zmq driver and not something
that is really able to be abstracted away. This change moves the key
generation bits to the zmq driver and uses the server / client construct
to generate the keys and exit. Documentation and bootstrap processes
have all been updated to support this new layout. Driver code has also
been cleaned up as there's no need for the key generation bits to exist
in the base class.

Signed-off-by: Kevin Carter <kevin@cloudnull.com>