Add server side logic to mark nodes active
When a node is active, work will be sent, if a node is inactive no work
will be sent to the node and that node will not be removed when expired.
Upon the nodes next healthcheck, the node will be reset as active and
all queued tasks will be sent at that time.

Related-Bug: #343
Signed-off-by: Kevin Carter <kevin@cloudnull.com>