Allow drivers to run in isolation
This change will allow each driver to run in isolation and ensure that
the options provided by the driver are only present in CLI interactions
when the driver would be functional. Non-functional drivers will be
omitted in the help output. This is a UX improvement, as well as a
packaging one.

Signed-off-by: Kevin Carter <kecarter@redhat.com>