rename zmq driver to zeromq
the zmq driver has been renamed to correct an import issue with the zmq
module using the same namespace as the zmq driver. While this issue can
be "fixed" with absolute imports, the driver should never have been
using this same namespace and since zmq is no longer the default driver
now is the time to correct this mistake.

Signed-off-by: Kevin Carter <kecarter@redhat.com>