Allow drivers to run in isolation (#380)
* Allow drivers to run in isolation

This change will allow each driver to run in isolation and ensure that
the options provided by the driver are only present in CLI interactions
when the driver would be functional. Non-functional drivers will be
omitted in the help output. This is a UX improvement, as well as a
packaging one.

Signed-off-by: Kevin Carter <kecarter@redhat.com>

* rename zmq driver to zeromq

the zmq driver has been renamed to correct an import issue with the zmq
module using the same namespace as the zmq driver. While this issue can
be "fixed" with absolute imports, the driver should never have been
using this same namespace and since zmq is no longer the default driver
now is the time to correct this mistake.

Signed-off-by: Kevin Carter <kecarter@redhat.com>